requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.0.8
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
google-genai==0.2.2
python-dotenv==1.0.1
